Replace CallX2 with latest feature version

+++ CallX2/feature-reels/src/main/java/com/callx/app/feed/controllers/ReelUiController.java
@@ -1602,7 +1602,7 @@ public class ReelUiController {
         reelPinnedCommentContainer.setVisibility(View.VISIBLE);
         if (tvPinnedAuthor != null) tvPinnedAuthor.setText(reel.pinnedCommentAuthorName);
         if (tvPinnedText != null) tvPinnedText.setText(reel.pinnedCommentText);
-        if (tvPinnedLikes != null) tvPinnedLikes.setText(delegate.formatCount(reel.pinnedCommentLikes));
+        if (tvPinnedLikes != null) tvPinnedLikes.setText(delegate.formatCount((int) reel.pinnedCommentLikes));
         if (ivPinnedAvatar != null && delegate.isAdded()) {
             Glide.with(delegate.requireContext())
                 .load(reel.pinnedCommentAuthorAvatar)
